Move slicesources tests to tests folder
