+++ volumina/pixelpipeline/slicesources.py
@@ -261,63 +261,3 @@ class SyncedSliceSources(QObject):
         for i in range(len(self._through)):
             through[self._sync_along[i]] = self._through[i]
         sliceSrc.through = through
-
-
-import unittest as ut
-
-# *******************************************************************************
-# S l i c e S o u r c e T e s t                                                *
-# *******************************************************************************
-
-
-class PlanarSliceSourceTest(ut.TestCase):
-    def setUp(self):
-        import numpy as np
-        from .datasources import ArraySource
-
-        self.raw = np.random.randint(0, 100, (10, 3, 3, 128, 3))
-        self.a = ArraySource(self.raw)
-        self.ss = PlanarSliceSource(self.a, projectionAlongTZC)
-
-    def testRequest(self):
-        self.ss.setThrough(0, 1)
-        self.ss.setThrough(2, 2)
-        self.ss.setThrough(1, 127)
-
-        sl = self.ss.request((slice(None), slice(None))).wait()
-        self.assertTrue(np.all(sl == self.raw[1, :, :, 127, 2]))
-
-        sl_bounded = self.ss.request((slice(0, 3), slice(1, None))).wait()
-        self.assertTrue(np.all(sl_bounded == self.raw[1, 0:3, 1:, 127, 2]))
-
-    def testDirtynessPropagation(self):
-        self.ss.setThrough(0, 1)
-        self.ss.setThrough(2, 2)
-        self.ss.setThrough(1, 127)
-
-        self.triggered = False
-
-        def check1(dirty_area):
-            self.triggered = True
-            self.assertEqual(dirty_area, sl[:, 1:2])
-
-        self.ss.isDirty.connect(check1)
-        self.a.setDirty(sl[1:2, :, 1:2, 127:128, 2:3])
-        self.ss.isDirty.disconnect(check1)
-        self.assertTrue(self.triggered)
-        del self.triggered
-
-        def check2(dirty_area):
-            assert False
-
-        self.ss.isDirty.connect(check2)
-        self.a.setDirty(sl[1:2, :, 1:2, 127:128, 3:4])
-        self.ss.isDirty.disconnect(check2)
-
-
-# *******************************************************************************
-# i f   _ _ n a m e _ _   = =   " _ _ m a i n _ _ "                            *
-# *******************************************************************************
-
-if __name__ == "__main__":
-    ut.main()
